whoscored/requirements.txt
beautifulsoup4
selenium
webdriver-manager
pyyaml
black